Repair all problems for IoTDBAlterTimeSeriesTypeIT

+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/storageengine/dataregion/DataRegion.java
@@ -3091,6 +3091,19 @@ public class DataRegion implements IDataRegionForQuery {
     Set<ModificationFile> involvedModificationFiles = new HashSet<>();
     List<TsFileResource> deletedByMods = new ArrayList<>();
     List<TsFileResource> deletedByFiles = new ArrayList<>();
+    boolean isDropMeasurementExist = false;
+    boolean isDropTagExist = false;
+
+    if (deletion instanceof TableDeletionEntry) {
+      TableDeletionEntry entry = (TableDeletionEntry) deletion;
+      isDropMeasurementExist = !entry.getPredicate().getMeasurementNames().isEmpty();
+    } else {
+      TreeDeletionEntry entry = (TreeDeletionEntry) deletion;
+      if (entry.getPathPattern() instanceof MeasurementPath) {
+        Map<String, String> tagMap = ((MeasurementPath) entry.getPathPattern()).getTagMap();
+        isDropTagExist = (tagMap != null) && !tagMap.isEmpty();
+      }
+    }
     for (TsFileResource sealedTsFile : sealedTsFiles) {
       if (canSkipDelete(sealedTsFile, deletion)) {
         continue;
@@ -3187,7 +3200,7 @@ public class DataRegion implements IDataRegionForQuery {
       } // else do nothing
     }
 
-    if (!deletedByFiles.isEmpty()) {
+    if (!deletedByFiles.isEmpty() && !isDropMeasurementExist && !isDropTagExist) {
       deleteTsFileCompletely(deletedByFiles);
       if (logger.isDebugEnabled()) {
         logger.debug(
